pyproject.toml
[tool.poetry]
name = "smart_queue_system"
version = "0.1.0"
description = ""
authors = ["Jean Dossa Tsimbazafy <493354@muni.cz>"]

[tool.poetry.dependencies]
python = "^3.9"
pendulum = "^2.1.2"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
drf-spectacular = {extras = ["sidecar"], version = "^0.21.2"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
